Added the client side SimEthereal support and removed the naive listeners.  This required some slight reworking of the model management in ModelViewState because objects may not come and go from view even if the owning player is still online.  Note: this is one of the cases where an ES would have made things simpler. Random fixes and tweaks to get everything working. This is still just doing raw updates without interpolation... that'll be next.  But at least the networking is efficient now.

+++ sim-eth-basic/src/main/java/example/sim/SimplePhysics.java
@@ -89,9 +89,9 @@ public class SimplePhysics extends AbstractGameSystem {
         return result.bodyId.intValue();       
     }
     
-    public boolean removeBody( int id ) {
+    public boolean removeBody( long id ) {
         Body body = index.remove(id);
-        if( body == null ) {
+        if( body == null ) {        
             return false;
         }
         toRemove.add(body);
